Check off from list when user is near marker
This includes a change to map width because it previously covered the whole screen. Google-map-markers were rearranged to match the order of paper-checkbox.

This design does not follow good Polymer idioms, but it does show how this feature could be implemented. Future work will be to make custom checkboxes / markers that can manage this behavior themselves (i.e. more of an object-oriented approach). For now, we're taking it as some light technical debt.