Change when bundled files are moved
Previously the output files were moved to 'out' in the doCompile function.
However, when working off master, this meant that the script would fail
because 'out' was not yet created---it is created as part of cloning
gh-pages. Now, the moving of the bundled files has been pulled out of
doCompile and is done after that function's invocation, only after
we know we're doing a deploy (i.e. we're not just checking if the build
fails or not).