Refactor map-view
This includes two major changes: adding `proximity-checkbox` and using a
data object.

`proximity-checkbox` is a new element that checks itself when the user
gets near. Like the previous implementation, it is "sticky" in that once
checked, it cannot be unchecked. Unit tests verify its behavior, which is
important since--as of this commit--there is no way to spoof location.

Both `proximity-checkbox` and `google-map-marker` needed location information,
so to remove the redundancy, I introduced a `locations` data element
to `map-view`. The `dom-repeat` templates allow you to stamp out different
elements per array item, which is now done for both the markers and the
checkboxes.