Use vh to specify map height
vh is a unit where 1vh = 1% of the screen height. Although 50% did not
work for us (Dakota and me), vh did.

Right now the whole block is scrollable, which is not what we want,
but we think that behavior is coming from the paper-tabs behavior.
Once we take that away, we can revisit this to get better dynamic
screen behavior.