Use app-route for routing (#12)
* Use app-route for routing

The previous approach used a shared `page` property for routing, which
was very simple, but it made it hard to get to specific pages. To
debug, say, the map page, someone would have to click through each
page. This approach also would eventually fail when we tried to handle
page reloading, as far as I can tell.

The new approach uses the app-route element and its related
app-location element. Documentation on these can be found at
https://www.polymer-project.org/1.0/toolbox/routing. The crux of it is
that app-location pulls the current address from the browser and
app-route parses it into a useful JSON object. The new `routeData`
property---now shared among views the way that `page` used to
be---represents the application's current route. The last part of the
URL indicates what view to load, so we can hop directly to the map
page by putting "map" as the route, for example.

This approach has the downside that someone could modify the URL to
get to a part of the application that doesn't make any sense yet, so
in the future we'll have to add some logic to prevent that. In the
meantime, it feels to me like this approach is easier for us to
debug. Also, because of how this this routing is now fully
encapsulated in view-behavior, especially with the new `_goToPage`
function, we should be able to easily swap to another routing scheme
if we need to later.

* Handle 404

The app will redirect to the new missing-view if it hits a 404. That is,
this will happen if the user enters a URL to a page that doesn't exist
(or, I suppose, if we accidentally botch a link).