Replace hot-spot-id with current-site (#19)
* Rename hot-spot-id to current-location

* Remove implementation tests

Several of these tests were looking at the implementation, not the
behavior. That's dangerous, so I removed them.

* Revise to return location objects

This also includes a complete overhaul of the unit tests to behavior-driven
development (bdd) instead of tdd-style. I think it reads a bit better, and
it might help others read the tests and consider the distinction between
unit testing of features vs testing of implementations.

* Rename location to site

The word 'location' was overloaded in the implementation, so now this
uses 'site' to mean a geolocative landmark that is significant to play.

* Show current site on map view

This is not a final feature but demonstrates the integration of
the current-site element data.