Add starter test for checklist-view
This test demonstrates how we can start to describe the desired
behavior of checklist-view. All it says right now is that when the
first checkbox is checked, then the responses array contains that
element. This is very simple, but the next step will be to ensure that
it is removed when it is unchecked, and from there, that this works
robustly with something like check1, check2, uncheck1 --> result is
only item '2' is returned.

Getting this to work required bringing in iron-test-helpers for the
MockInteractions (that was new to me!) as well as using flush to
operate a unit test involving a DOM-modifying element.  For examples
of the former, I turned to the unit tests around paper-checkbox
itself, which can be found at
https://github.com/PolymerElements/paper-checkbox/blob/master/test/basic.html.
For details on the latter, see
https://www.polymer-project.org/1.0/docs/tools/tests.