Revert to dynamic loading of title-view
My idea to load the title-view automatically, referenced in #23, turns
out to not have been a good idea. My previous commit broke the CI build
with some repeat-definition errors, all related to my trying to outsmart
the PRPL pattern system. The moral of the story: premature optimization is
the root of all programming evils. If it turns to be a problem, we can
go back and look at synchronously-loading this resource, but I doubt it
will come to that.