Control developer mode ("devmode") through URL
Developer mode is enabled by loading any page with a "/d" at the end
of the url, such as with `http://localhost:8080/#/title/d`. Once you
set this, it "sticks" even though the extra flag disappears as you
navigate. The only way to disable developer mode is to reload a page
manually while omitting the /d flag.

Developer mode is currently integrated into two views:

- `timer-view` adds a skip button

- `map-view` hides the "use current location" and "override location"
  buttons unless dev mode is enabled

I have added a ".debugging" class to the `shared-styles`
stylesheet. All of the content that is part of a dev-mode-only element
can implement this class and be displayed in a lovely light magenta.

Note that the behavior of `map view`'s debugging features haven't
changed, although they could use some work---making location override
modal and turning off GPS location while it's on, for example. That's
not part of this commit, however.