Show only instructions or arrival div
Making this work required changing current-site so that it returned null
instead of undefined when we are not at a site. The implementation of current-site
treated undefined as if it were a keyword, but it's not---it's a non-value,
unlike null, which is a real value with... no value? Javascript is weird about
this, but I needed some value in order for the hidden attributes to work properly.