Restrict arrival message to unvisited sites
Working on this makes me see an opportunity for refactoring, where we have
multiple locations that make assumptions about the structure and
behavior of the map data. For example, the fact that each location has
to have a 'visited:false' flag is required for this test to work---and
indeed for the element to work---but that idea is spread throughout the
code. We should come up with a way to remove that DRY violation.