Fix template declaration
It looks like there was an errant find-and-replace that was committed,
but it has been fixed, which I believe fixes #50.

As part of the refactoring while in these files,
current-site_test.html was moved to the top level of the test
directory, bringing it in line with the rest of the test cases. Also,
the `before` methods were changed to `beforeEach`, which didn't change
their behavior (the tests still pass the same) but certainly makes
them more authentic to what was intended.