Refactor to remove unsupported JS language level
The previous implementation used the 'of' keyword, which is not supported
at the language level of the Polymer build system.

Parts of this file were also unconventionally arranged, so I cleaned it
up by running Visual Studio Code's auto-formatter, which I have found to
be more reliable for Polymer than Atom's.