Fix css height (#62)
* Fix height to extend to 100% of sreen size.

* Replace heading styles with divs

The h1 and h2 tags that were here had been here a while, and they were
originally used as shorthand by someone to mean 'make this text large.'
However, this text never semantically matched the concept of a heading,
and when we got to styling, we got bit by the fact that heading types
come with custom margins. Now, we are using divs instead, with a class
that specifies what they are for.

It would be possible to either promote the class to shared-styles or come
up with some kind of 'custom-paragraph' element that encapsulates the
idea of 'this text is to be shown to the user.' That's beyond my current
scope though.

* Reword intro message

The original sentence was stilted and used the word 'around' twice. I
think this articulation is more active and readable.