Remove prairie-creek test case
The lone unit test in there has been superceded by the behavior of
supporting page reload with localstorage. This does mean there is an
opportunity for UX improvement: if someone loads the title page but
the game is in progress, we should probably route them to the map
view, from where they can reset if they want to.