Handle page reload gracefully
Implementation uses iron-localstorage. There is a funny little dance
around mapData to ensure the binding works right; it should probably
be extracted, but we're taking it as technical debt for now.

As part of this work, the tests for prairie-creek-game were removed.
The lone unit test in there has been superceded by the behavior of
supporting page reload with localstorage. This does mean there is an
opportunity for UX improvement: if someone loads the title page but
the game is in progress, we should probably route them to the map
view, from where they can reset if they want to.