Fix spirit determination on non-Chrome browsers
As described in #72, Firefox was showing a different result for `event.target` than in Chrome: in the former, it was one of the subelements of `paper-checkbox`, but in the latter, it was the `paper-checkbox` itself. This was a particular problem here since the Firefox-referenced element did not have the `checked` field that its parent---the `paper-checkbox`---did. Referencing `event.currentTarget` in the tap listener instead gives reference to the `paper-checkbox` on all tested platforms, which means we can reliably get to the `checked` value of the checkbox. Hence, this fixes #72.