Disable gps ping except on map view and on first check
The first check is needed to ensure that we can get into the map
in a good state, and the fact that we do this check is baked into
current-location.

For other checks, I've added a listener to iron-pages that gets
notified after a selection is made. Note that this is separate from
all the plumbing required for PRPL routing: this new on-iron-select
event is fired _after_ the selection process is done, so that we can
use this.$.pages.selectedItem to get the currently-selected element.
We only turn gps pinging on if that element has the gps-active
attribute.

This fixes #69