Make button text dynamic
The previous text referred to a 'checklist', which is not part of the
player's nomenclature for the game. Simply making the button appear when
it is time to click it was a bad UX option (backed by the theory that
we should not change a UI on the fly). I found satisfaction with updating
the button text dynamically based on the state of the timer. One could
extend this to make the text reflect the sense being used, but that
seemed out of scope for this fix.

This fixes #77.