Prevent unwanted gps update off map screen
There was a potential timing defect previously:
1. user-location is active, so it sends an async request for gps info
2. user-location becomes inactive
3. asyc call returns, and so position is updated.

This commit fixes that problem, by checking the active status flag
again just before updating latitude and longitude. NB: It has to be
checked before the call, too, to prevent the expensive GPS check.

The _first property was added because we need to grab the GPS position
the first time we load this element, regardless of whether it is
active, so that we have some data to base the rest of our logic upon.