Add bigtext CSS class
The text style was being used and overridden in an ad hoc basis in
several places, but this made it hard to maintain. We introduced
"bigtext" for text that should be a bit larger. This is not quite
perfect yet, but it's a more maintainable structure. Further work is
needed, like in this commit where we have spacing around a div and
font styles only on the text class.