Prevent game start without geolocation
The play button on title view is disabled by default, but it becomes
enabled when geolocation is available. This required marking
`title-view` as `gps-active` within `prairie-creek-game`, which
probably has served to fix some of our problems as well. My local
testing indicates that this fixes #107.

The 'Please enable GPS' message is no longer included at all in
map-view, since this is now prevented one level up, in a good example
of the HCI principle: if you don't want someone to access something in
the wrong state, make it impossible.