Prevent game start without geolocation (#108)
* Include enhanced developer mode options through drop-menu

"Enhanced developer mode" separates different devmode features for
more fine-grained debugging. Note that the ability to enter dev mode
with the '/d' custom URL has been removed; instead, all of the options
are accessed through the drop menu's new "Advanced" submenu.

The enhanced developer mode options include:

+ echo: Show latitude, longitude, and gps error messages in a text
widget at the top of the screen.

+ warp: Enable or disable location warp (teleporting to sites)

+ override: Enable or disable location override (tap to set location)

Note that with the removal of the timer-based position acquisition, we
no longer needed a way to disable GPS features in developer mode. This
was a solution to an artificial problem, that we were asking for
location on the timer. Now, if you're debugging at a desktop, you can just
set your location with override or sensor manipulation without worrying
about warping back to your home location. The only thing that is lost
is the ability to warp to locations on site in a "sticky" way: sensor
changes will take you out of that site. Realistically, though, that's
not expected gameplay behavior, so I am not worried about that for the
purpose of these developmer mode options.

* Prevent game start without geolocation

The play button on title view is disabled by default, but it becomes
enabled when geolocation is available. This required marking
`title-view` as `gps-active` within `prairie-creek-game`, which
probably has served to fix some of our problems as well. My local
testing indicates that this fixes #107.

The 'Please enable GPS' message is no longer included at all in
map-view, since this is now prevented one level up, in a good example
of the HCI principle: if you don't want someone to access something in
the wrong state, make it impossible.