Address lint warnings (#110)
I did not know that behaviors had to be marked up with special
metadata so that the linter would read them properly; now that this is
in place, we should not get linter warnings regarding ViewBehavior any
more. I found out about this by looking at
https://github.com/PolymerLabs/polylint/issues/130.

The other changes were pedestrian fixes.