Spirit naming ux (#111)
* Add paper-input for spirit naming

* Clear name input after leaving page

* Store name input and display on team-view

* Add default placeholder names for spirits

Restructured the nature and artificial items in map-data to contain both the text and the name for the spirit. The text has also been updated. Updated spirit-view and team-view to work with the new format. Name input starts out with default name in it.

* Reorganize spirit-naming-us

Move the spirit-view text prompt to the top of the view and resolve blank name to remain as defaul