Allow timer manipulation through drop-menu
There is now an option in the drop-menu's advanced options to use a
short timer, which is set to two seconds. The skip feature---which was
introduced for debugging early in development---has been removed for
now.

A new test was added to ensure that a shorter timer is used when
short-timer is true.

This fixes #114, which was a remnant of the transition to enhanced dev
mode, an oversight in the original implementation.