Add map-data asset validation tests (#122)
There are now tests for the icons, gray icons, and spirit images, so
that the build itself will fail if these are not in place.

Simply adding the tests caused the build to appropriately fail, since
prior to this, there were missing images, relating to the revision of
the sites. I have renamed the spirit images from the old sites to the
new sites, without any consideration of their content---this was just
to get the tests passing and, of course, the game working.