Cache spirit images
This loads all the spirit images during app initialization.  Their
paths are stored in the cache, but they're never actually read: simply
holding the persistent reference to them is enough that the browser
keeps the referenced images in memory, so that when other places go
looking for them by path (e.g. spirit-view), there they are.

Fixes #119