Handle back button (#127)
We have implemented a state machine using appState, which is defined in `prairie-creek-game` and bound into all views (i.e. elements that use `ViewBehavior`). This replaces the `app-route` approach, which was robust neither in the face of back button usage nor page reload.

`currentSite` is now being stored in localstorage as well in order to support page reload. The only slightly weird behavior we saw was that reloading the timer-view resets the timer, but that's an acceptable loss for the elegance of the overall solution.

Note that reloading a page that uses currentSite may not behave right if you are using dev mode. This is because reloading the page triggers a refresh of latitude/longitude, which updates currentSite, etc. You can still check the behavior locally by overriding your browser's geolocation sensors.