Show link to map when outside Prairie Creek Park
Now, when devmode's warp feature is enabled, you can warp to the park via the title screen, whereas it used to only be in the map view. This allows us to test the application from elsewhere.

The tests for title-view have been changed because there was no longer a need to have a disabled play button; this feature subsumes that.